backend/cross-chain-service/xmrt-cross-chain-service/src/routes/wormhole.py
from flask import Blueprint, jsonify, request
import os
from dotenv import load_dotenv
import json
from datetime import datetime

load_dotenv()

wormhole_bp = Blueprint('wormhole', __name__)

# Wormhole configuration
WORMHOLE_CORE_BRIDGE = os.getenv('WORMHOLE_CORE_BRIDGE', '0x706abc4E45D419950511e474C7B9Ed348A4a716c')
WORMHOLE_TOKEN_BRIDGE = os.getenv('WORMHOLE_TOKEN_BRIDGE', '0xF890982f9310df57d00f659cf4fd87e65adEd8d7')

# Supported chains for Wormhole
SUPPORTED_CHAINS = {
    'ethereum': {'chain_id': 2, 'name': 'Ethereum'},
    'polygon': {'chain_id': 5, 'name': 'Polygon'},
    'bsc': {'chain_id': 4, 'name': 'BSC'},
    'avalanche': {'chain_id': 6, 'name': 'Avalanche'},
    'arbitrum': {'chain_id': 23, 'name': 'Arbitrum'},
    'optimism': {'chain_id': 24, 'name': 'Optimism'}
}

@wormhole_bp.route('/supported-chains', methods=['GET'])
def get_supported_chains():
    """Get list of supported chains for Wormhole bridge"""
    try:
        return jsonify({
            'success': True,
            'data': {
                'supported_chains': SUPPORTED_CHAINS,
                'total_chains': len(SUPPORTED_CHAINS),
                'bridge_type': 'wormhole'
            }
        })
    except Exception as e:
        return jsonify({
            'success': False,
            'error': str(e)
        }), 500

@wormhole_bp.route('/estimate-fee', methods=['POST'])
def estimate_bridge_fee():
    """Estimate fee for Wormhole bridge operation"""
    try:
        data = request.get_json()
        source_chain = data.get('source_chain', '')
        target_chain = data.get('target_chain', '')
        amount = data.get('amount', 0)
        
        if not source_chain or not target_chain:
            return jsonify({
                'success': False,
                'error': 'Source and target chains are required'
            }), 400
        
        if source_chain not in SUPPORTED_CHAINS or target_chain not in SUPPORTED_CHAINS:
            return jsonify({
                'success': False,
                'error': 'Unsupported chain'
            }), 400
        
        # Simulate fee estimation (in production, this would call actual Wormhole APIs)
        base_fee = 0.005  # ETH
        amount_fee = float(amount) * 0.001  # 0.1% of amount
        total_fee = base_fee + amount_fee
        
        return jsonify({
            'success': True,
            'data': {
                'source_chain': source_chain,
                'target_chain': target_chain,
                'amount': amount,
                'estimated_fee_eth': total_fee,
                'estimated_fee_usd': total_fee * 3500,  # Approximate ETH price
                'estimated_time_minutes': 15,
                'bridge_type': 'wormhole',
                'timestamp': datetime.now().isoformat()
            }
        })
        
    except Exception as e:
        return jsonify({
            'success': False,
            'error': str(e)
        }), 500

@wormhole_bp.route('/initiate-transfer', methods=['POST'])
def initiate_transfer():
    """Initiate a Wormhole bridge transfer"""
    try:
        data = request.get_json()
        source_chain = data.get('source_chain', '')
        target_chain = data.get('target_chain', '')
        amount = data.get('amount', 0)
        recipient = data.get('recipient', '')
        
        if not all([source_chain, target_chain, amount, recipient]):
            return jsonify({
                'success': False,
                'error': 'All fields are required: source_chain, target_chain, amount, recipient'
            }), 400
        
        # Simulate transfer initiation (in production, this would interact with Wormhole contracts)
        transfer_id = f"wh_{int(datetime.now().timestamp())}"
        
        return jsonify({
            'success': True,
            'data': {
                'transfer_id': transfer_id,
                'source_chain': source_chain,
                'target_chain': target_chain,
                'amount': amount,
                'recipient': recipient,
                'status': 'initiated',
                'estimated_completion': datetime.now().isoformat(),
                'bridge_type': 'wormhole'
            }
        })
        
    except Exception as e:
        return jsonify({
            'success': False,
            'error': str(e)
        }), 500

@wormhole_bp.route('/transfer-status/<transfer_id>', methods=['GET'])
def get_transfer_status(transfer_id):
    """Get status of a Wormhole bridge transfer"""
    try:
        # Simulate transfer status (in production, this would query Wormhole APIs)
        status_options = ['initiated', 'pending', 'confirmed', 'completed']
        
        return jsonify({
            'success': True,
            'data': {
                'transfer_id': transfer_id,
                'status': 'completed',  # Simulate completed for demo
                'confirmations': 65,
                'required_confirmations': 65,
                'bridge_type': 'wormhole',
                'last_updated': datetime.now().isoformat()
            }
        })
        
    except Exception as e:
        return jsonify({
            'success': False,
            'error': str(e)
        }), 500

@wormhole_bp.route('/bridge-stats', methods=['GET'])
def get_bridge_stats():
    """Get Wormhole bridge statistics"""
    try:
        return jsonify({
            'success': True,
            'data': {
                'total_volume_24h': 1250000,  # USD
                'total_transfers_24h': 342,
                'average_fee_eth': 0.008,
                'average_completion_time_minutes': 12,
                'supported_chains_count': len(SUPPORTED_CHAINS),
                'bridge_type': 'wormhole',
                'last_updated': datetime.now().isoformat()
            }
        })
        
    except Exception as e:
        return jsonify({
            'success': False,
            'error': str(e)
        }), 500

@wormhole_bp.route('/health', methods=['GET'])
def health_check():
    """Health check for Wormhole service"""
    try:
        return jsonify({
            'success': True,
            'data': {
                'service': 'wormhole',
                'status': 'healthy',
                'version': '1.0.0',
                'supported_chains': len(SUPPORTED_CHAINS),
                'timestamp': datetime.now().isoformat()
            }
        })
        
    except Exception as e:
        return jsonify({
            'success': False,
            'error': str(e)
        }), 500

